treino de api

+++ MVC/exercicios/treino-api/NotaFiscal/DTO/NotaFiscalView.cs
@@ -7,7 +7,7 @@ namespace NotaFiscal.DTO
     public class NotaFiscalView
     {
         public int Id { get; set; }
-        public float Valor { get; set; }
+        public double Valor { get; set; }
         public DateTime DataEmissao { get; set; }
         public Cliente Cliente { get; set; }
         public ICollection<ProdutoNotaFiscal> ProdutosNotaFiscal { get; set; }
